refactor: Phase 1 - merge engine and config_manager dups
- Create core/ with EventBus, ApplicationState, ErrorInfo, EventEmitter
- Transform engine.py into deprecation compat layer
- Transform config_manager.py into deprecation compat layer
- Add service_container global instance to service_container.py
- config_manager provides local LLMConfig/CacheConfig/RetryConfig for compat
